Move string space functions to string_helpers
