+++ src/parse_helper.h
@@ -30,6 +30,3 @@ enum token_type char_to_token(char input);
 struct parser_context *init_parser(void);
 void destroy_parser_context(struct parser_context *context);
 
-int64_t find_first_nonspace(const char *string);
-int64_t find_last_nonspace(const char *string);
-int64_t strip_surrounding_space(const char *string, char **dest);
